Switched EasyCommands.Tests references for SE and MDK to NuGet. Checked in csproj files. (#2)
* Added the first ScriptTests along with Moq.

Added the first ScriptTests along with Moq mocking framework.
We can now write "end-to-end" tests where we pass the entire
sample script as input, and perform verifications on the
blocks (using Moq to mock the SE interfaces) as way of assertion.

I think there's a lot of cleanup we can do to the test style,
but I'll save that for a follow-up.

Note that as part of this I had to modify ParameterParsers and
ParameterProcessors to avoid double-initializing the respective
dictionaries.

I also added some delegates to the script to allow test authors to
customize the behavior of finding blocks, setting custom data, and
setting broadcast messages. (I only used the first two in this test).

* Added .editorconfig to .gitignore

* Added NuGet dependencies for SE SDK and MDK.

Added NuGet dependencies for SE SDK and MDK.
Changed gitignore to stop ignoring csproj files,
as now there are no local paths being referenced,
since all deps are being pulled through NuGet.

Also stopped ignoring .editorconfig.