Consolidated and improved Operations support
This commit moves the existing interface functions on Primitives to propery operations and introduces a better pattern for dealing with operating on different types
of Primitives.  It does so by working on the existing object types instead of relying on PrimitiveType.  This also enabled me to do better operation handling, such
as dealing with operations which work even if the types of a & b are swapped (addition, multiplication).

I was also able to model comparison, and, or and not as operations returning a boolean primitive.