Added "all" and implicit selector support
This change allows you to perform an action on all blocks of a given type without defining a group.

The new selectors can be used in block commands to perform actions, or in aggregate block conditions to check for block state.

You can also use this to get aggregate block properties.