Revamped ScriptTests by adding ScriptTest utility class
Revamped ScriptTests by adding ScriptTest utility class
This greatly reduces the required boilerplate in writing script tests
by delegating most of that to the construction of a ScriptTest instance.

ScriptTest authors only need to provide the actual script,
the mocked objects, and the verifications/assertions for those mocks.

As part of this, I replaced block and group entity providers with
mocking of the GridTerminalSystem. However, I haven't yet configured
it to work properly for display blocks. I wanted to get this bit out
and reviewed before then proceeding to address the display block gap.

As a result, you'll see changes to EasyCommands for the removal of the
block provider delegates, as those are no longer required.

Other notable changes:
* Changed STATE in Program to be public so we can both examine it
  and set it from tests
* Changed the BroadcastMessageProvider delegate to not be static
  so that we don't have to potentially reset it after each test run