Added target support to Camera and Sensor Blocks
This commit adds basic target support for Camera and Sensor Blocks
It also fixed a couple bugs with getting and setting custom properties which were discovered during testing.

Also adds block handler tests for Camera and Sensor Blocks along with a MockEntityUtility for adding commonly used mocks in the future.