Migrated RocketVolley to EasyCommands
This commit also fixes a bug with missing "not" processing for BlockCommands.

It also adds a few more keywords which should map to the GunBlockHandler.

Lastly, it adds another utility to mock expected actions in MockEntityUtility and sets the default function parse amount for tests to 1000, so that tests always
parse in 1 turn.

Deleted RocketVolley as it should no longer be needed