This commit adds the ability to process multiple possible parsings fo… (#57)
* This commit adds the ability to process multiple possible parsings for ambiguous requests such as selector + property, where it could either be an aggregate variable
or a block command depending on the context.

Implicit Aggregate Properties uses a new AggregationMode for "value".  This aggregation mode throws an exception if the input selector has more than 1 value.  The
result is that you will get an exception during execution if you try to use an implicit aggregation on a selector property that contains more than one block.

Also adds test and modifies some of the existing script tests to confirm the new parsing is working.

* Fixed ParameterProcessor Process Description Typo

Co-authored-by: Merlin <merlinofmines@gmail.com>