Migrated Ship Proximity Script to EasyCommands (#49)
This commit does a few things.  First, it added a new "targetVelocity" attribute to cameras and sensors for getting the target's velocity.  Returns Vector.Zero if no
target.

Second, it adds many trigonometric functions such as sin,cos,tan,asin,acos,atin.

Third, it adds round as a unioperand function to round a given number to the nearest integer.

Fourth, it adds the doct product bi-operand (".") to get the dot product between two vectors.

Fifth, it adds the ability to mock a detected entity including detected velocity to MockEntityUtilities.

Finally, using all the above, it adds a new ScriptTest showing a working script to calculate the distance and closest approach of a given target vessel.