Added line number to exception output when unable to parse
This commit adjusts the parsing to include the line number in the exception output when the command cannot be parsed, which drastically reduces the time spent
searching for the culprit in a script that isn't working.

As part of this commit, the parsing is changed to not pre-emptively ignore comments and empty space.  Instead, these entries are ignored while parsing the commands
so that the correct line numbers remain intact