Added new ValueProperty type and AMOUNT ValueProperty
This commit adds a new "ValueProperty" which is a property that has a variable value.  This is useful when there is a variable dimension to the given property.
Example: "gold ingot" amount, where "gold ingot" is variable dimension.

This commit also adds a skeleton for CargoBlockHandler which will leverage the new AMOUNT ValueProperty