Added Items parsing support and basic cargo container amount support
This commit puts in place a mapping between simple words and the core in game items so that we can start performing actions on inventory items.  The first of
these, implemented in this commit, is the ability to retrieve an amount of a given item type from a cargo container.

It also relaxes the parsing around ValueProperties so that you can provide the value before or after the value property.  This enables a statement like "assign a
to the amount of "gold ingot" in the "cargo container"

You can also select multiple item types at a time using comma separated values, such as "iron ore, iron ingot"