Inventory support enhancement (#77)
* Added PropertySupplier and changed BlockHandlers to inject PropertySupplier to all handlers

* Added new ValueProperty type and AMOUNT ValueProperty

This commit adds a new "ValueProperty" which is a property that has a variable value.  This is useful when there is a variable dimension to the given property.
Example: "gold ingot" amount, where "gold ingot" is variable dimension.

This commit also adds a skeleton for CargoBlockHandler which will leverage the new AMOUNT ValueProperty

* Added Items parsing support and basic cargo container amount support

This commit puts in place a mapping between simple words and the core in game items so that we can start performing actions on inventory items.  The first of
these, implemented in this commit, is the ability to retrieve an amount of a given item type from a cargo container.

It also relaxes the parsing around ValueProperties so that you can provide the value before or after the value property.  This enables a statement like "assign a
to the amount of "gold ingot" in the "cargo container"

You can also select multiple item types at a time using comma separated values, such as "iron ore, iron ingot"

* Added basic Transfer Command Support

This commit adds a TransferCommand for transferring items between containers.  It also fixed some issues with entity ids and implemented a mock utility for mocking
inventories.

* Added PropertyHandlers to CargoHandler for Ratio (%Volume), Range (Max Volume) and Volume (Current Volume).

Also Added tests around CargoBlockHandlers.
Also added "capacity" as a Range property keyword