Added Automated Production Support
This commit adds an AssemblerBlockHandler with the ability to automatically produce items.
This commit introduces blueprint items, though the words for inventory items are the same as for blueprints.
It also adds new "Produce" and "Destroy" ValueProperties, and changed the existing Property to "Supply"
I also removed a couple unnecessary variables in BlockHandler that weren't doing anything.
Also added mock utilities for BluePrints and Production Queues to MockEntityUtility