Added Ejector and Connector Block Handlers
This commit adds some basic support for ejectors and connectors.  It also extends properties on the Connector Block Handler to include all those if ejector, and also
Strength.

Notably, this commit also fixes a bug with Action where "stop collecting" and similar wouldn't work.  This was fixed by making "stop" a NotCommandParameter instead
of false and allowing Action to proceed if Not only is present (so that "stop the x" still works).