Added Hinge Block Handler
This commit separates rotors from hinges so that you can operate on the two indepedently, even if there are rotors and hinges in the same group.
It also adds several more properties to rotors/hinges including connect (attach), lock, and strength (torque).

Added torque as a new strength property word.