Changed List Separator Parsing Logic
This commit updates list separator parsing to treat comma delimited values as separate entities.
This is much better for dealing with multi-dimentional arrays.  With the + operation available to convert anything to a list with 2 characters
this is a small sacrifice to pay for much easier and cleaner multi-dimensional array support.

Removed now unused ListVariable as a bonus