Added Merge support for KeyedList
This commit adds support for merging two keyed lists, where all non-keyed values are kept and all keyed values are kept but de-duped.  In the case of de-dupes, the
second list values are kept.  Order is maintained as close as possible, where all keys remain in expected order except that duplicate keyed values from the first
list are removed.