Fixed bug with vectors being processed incorrectly
This commit fixes a previous bug where vectors (especially with negative values) would not parse correctly.
This commit also consolidates parsing of primitive values from strings to a single place and removes the overcomplicated negative parsing logic.
Lastly, I was able to clean up the NumericCommandParameter type entirely as a result of this change.