Fixed bug where single list values could not be added to lists
This commit fixes a bad assumption earlier that any single-item list should be converted to value instead of kept as a list.
The result was that you couldn't add single values to a list, even after wrapping the value with a list.
This commit fixes this assumption by moving the logic to convert single value lists into ListIndexVariable, where it belongs.
The result is that you can now add lists together properly, even with single values.  All other functionality is unchanged.