Fixed bug where a list of a single list was incorrectly resolved
This commit fixes an issue where a list of a single item, where the single item is a list, was not resolved correctly (treated like an index).  This was resolved
by creating a new concept for IndexVariable and explicitly using this only in cases where index resolution is expected.

I also added a previously missing parsing rule for treating a variable as a list (variable followed by list).

Additionally, fixed and issue where "none of" returns false for an empty list incorrectly by fixing some previously bad logic around what constitutes "matching".