Fixed Dynamic Property Resolution and added Properties Support
This commit fixes previously incorrect logic for fetching dynamic properties from IMyTerminalBlocks by propertyId.  It also adds support for String and long typed
properties.

This commit also adds a new "properties" property for all TerminalBlock so that all properties of a given block can be listed out (in order to be retrieved or
set) later.  This is handy if you aren't sure what all you are able to get/set on a block and want to inspect it first.

Lastly, this commit removes some boilerplate code for Property Words, which helps offset the added characters of this new functionality