Fixed bugs with incorrect comment parsing
This commit fixes a few bugs with comment parsing, including attempting to parse comments as colors as not parsing
functions with comments in weird places (like incorrect indentation and at the end of functions).

In addition this commit cleans up how line numbers are calculated while also reducing characters needed.