Separated increment from direction and added better increment support
This commit removes a dependency on increment from directions.  By doing so we can now do things like "decrease the upper limit", something that was impossible
before.  This change also enables us to increment without a value, such as "increase the rotor angle".  Additionally, this change does a better job of mapping the
get/set methods of SimplePropertyHandler correctly when directions are present.  This was also needed to enable behavior like "increase the upper limit".