Added reversed and sorted list support. Fixed bugs with my selector and range operation.
This commit adds support for reversing a given list using the "reversed" keyword (in front of requested list).  Similarly, it adds support for asking for the
sorted version of a given list.  The two can also be used in combination ("reversed sorted list").

Additionally, this commit extends the Range operation so that if the range given is (x..y) where x > y, the result is y, y-1, y-2, ..., x.  This also enables you
to reverse a list by requesting the range keys in reverse order (count of n - 1 .. 0)

Lastly, this commit fixes a recently introduced bug where setting "my display text to <x>" would set display text of all programmable blocks.  Whoops!