Consolidated DataFetchers and DataProcessors
This commit gets rid of redundant DataFetchers & DataProcessors.  As an added benefit, this change gets rid of a previous issue where arguments passed to delegates
would have incorrect values due to just in time resolution.  Furthermore, a large redunction in ".GetValue()" was removed with this, as now the convert functions
take the direct values from the data processors, instead of being passed the Data Processor.