Added support for Break, Continue and Return control commands
This command adds support for breaking out of and continuing through for each and conditional commands.  It also adds basic return support for returning from
functions without waiting for the entire execution.

Return values are not supported as this is an entirely different beast and very difficult to address.  Since variables set within a function are visible to the
caller, the workaround is simply to set a variable with a desired name and then reference it from the caller.

This commit also condenses the ControlCommand characters by getting rid of the painful switch statement and unnecessary ControlType construct.