ShipController Improvements (#143)
First, this commit adds a new handler for Cryo Chambers, which extend from Cockpits.  Second, this commit adds a new "Use" property that can be used to get whether
crockpits, cryo chambers, and/or remote controls are currently in use (occupied or being controlled).  Third, this commit fixes some issues with input and rotation
properties to make the default input forwards and scale rotational input to -1, 1.