Fixed Bug with Invalid Property Lookup
This commit fixes a previous bug where invalid property lookups would not present a reasonable error message to users.

This commit also removes debug output, which is not turning out to be useful in practice.  Easier and better debugging steps are already present in the debugging
guide, and it's easier to print out variable values/statements than trying to debug program output, anyway.