Added support for Dynamic Function Execution (#191)
This commit adds the ability to invoke functions dynamically using variables.  This is a powerful feature, but comes at the cost of compile time safety.

This commit moves the resolution of the function name to invoke to a runtime call.  It also leaves the resolution of parameter key names to runtime.  The existing error messages have been improved, but will now show up as runtime exceptions.

Lastly, this commit introduces a "Utilities" section to contain some useful utilities for saving characters on Type and Function declarations