Changed String Primitive Parsing
This commit changes string parsing so that anything wrapped in double quotes is not automatically converted to primitives.  Direct values can still be used, as can
conversions from string to the requested formats.

Also removed PrimitiveProcessor which is no longer needed.

Also redid some of the VectorVariable utilities to remove some now unnecessary logic.

Also added utility for Vector and moved utilities to the Utilities file.