Added UniOperand Cast and Type Operand Support
This commit updates the cast operation to act as either a UniOperand or a BiOperand.  As a UniOperand it will attempt to parse a string as the appropriate primitive
without needing to specify the expected type.  This is useful for parsing ambiguous input without causing script halting exceptions.

This commit also implements the Type operation, which enables you to get the type of a variable.  This is also useful for inspecting resolved input to verify it is
the correct type before attempting to use it as the type you are expecting.

This commit also fixes a previous bug where invalid cast and round operations were not logged correctly.

This commit also updates some incorrect documentation around vectors and adds documentation around parsing GPS coordinates.