Added HeatVent Block Handler
This commit adds support for HeatVents, a new block type introduced in the Warfare 2 update.  It supports all of the important properties for the new block type
using direct property access.

This commit also updates EasyCommands to a new version of SpaceEngineers Nuget dependency which has the new block types