Added Grid Block Handler (#236)
This commit adds basic support for Grids, providing access to get/set the grid name, and read-only access to whether the grid is static and the grid size
(large/small).

In addition, this commit removes the default String "name" property, which wasn't providing much benefit and usually caused confusion.  The new expectation is
that if you want to set a block's name, you must specify the "name" property.