Refactored PropertySupplier to support lists of properties
This commit changes PropertySupplier to accept a list of property values, as a pre-requisite to adding multi-property support.  This commit also changes the parsing
rules for selectors & properties so that properties & directions are bound to the closest selector while maintaining separation through a new
CommandSeparatorParameter which extracts out "to"

This commit also consolidates and re-organizes properties & value properties.