Fixed Implicit Aggregation Support
This commit changes the default aggregation mode to use a new special default aggregation instead of a sum.  This aggregation will return an empty list for an empty
selector, or the selector property directly for a selector of 1 block.  If the selector has multiple blocks, the aggregator will return the sum of the values if all
are numeric; otherwise will return a list containing the property values for each selector block.

This commit also consolidates the "name" and "names" properties, as this commit removes the unnecessary (and confusing) distinction between the two properties.