Add ProductionBlockHandlers.cs
This file provides handler class for IMyProductionBlock, the parental class for IMyAssembler and IMyRefinery blocks.  It also subsumes much of the code that had been present in AssemblerBlockHandler.cs, and as such that file needs to be removed from the repository.  This file does re-provide handler class for IMyAssemblerBlock and its properties and methods available above and beyond IMyProductionBlock.